Port werapi to 0.3
