+++ Cargo.toml
@@ -167,6 +167,7 @@ urlhist = []
 urlmon = []
 usbspec = []
 vsserror = []
+werapi = []
 winbase = []
 wincodec = []
 wincon = []
