--- src/werapi.rs
@@ -1,12 +1,0 @@
-// Copyright © 2015-2017 winapi-rs developers
-// Licensed under the Apache License, Version 2.0
-// <LICENSE-APACHE or http://www.apache.org/licenses/LICENSE-2.0> or the MIT license
-// <LICENSE-MIT or http://opensource.org/licenses/MIT>, at your option.
-// All files in the project carrying such notice may not be copied, modified, or distributed
-// except according to those terms.
-//! Function prototypes for Windows Error Reporting (WER)
-ENUM!{enum WER_REGISTER_FILE_TYPE {
-    WerRegFileTypeUserDocument = 1,
-    WerRegFileTypeOther = 2,
-    WerRegFileTypeMax,
-}}
